requirements.txt
praw~=7.1.0
python-dotenv~=0.15.0
pandas~=1.1.3